Add Tomi Manninen OH2BNS's axwrapper program.
Axwrapper first creates a pipe and then forks and execs the program
<server-program> with arguments given at the axwrapper command line.
The argv[0] argument is mandatory; further arguments are optional. The
parent process then sits and waits for any I/O to and from the user and
converts any carriage return characters from the user to line feeds and
any line feeds from the program to carriage returns.

This is useful when starting non-AX.25-aware programs from ax25d.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>