Substitute proper paths in ax25d.conf.
Depending on the prefix configured, the default ax25d.conf file may
contain incorrects paths.  Generate config file using the correct
sbin dir.

Assume node which is not part of ax25-tools to be installed with the
same prefix.  Don't probe - chances are the sysadmin of a node system
is going to install it after ax25-tools so we'd likely have to guess
which never is a good thing.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>